+++ CHANGES.rst
@@ -4,8 +4,8 @@ Changelog
 1.0.0rc2 (unreleased)
 ---------------------
 
-- Nothing changed yet.
-
+- Added batching module.
+  [sgeulette]
 
 1.0.0rc1 (2024-02-08)
 ---------------------